Method to method object
